requirements.txt
fastapi
python-dotenv
uvicorn[standard]
itsdangerous
aiosqlite
sqlalchemy
anthropic
python-telegram-bot>=21
httpx>=0.27
groq>=0.13
trafilatura>=2.0
lxml-html-clean
pydantic>=2
pyyaml
watchdog
apscheduler==3.10.*
playwright
pytest
pytest-asyncio
pymupdf
aiosmtplib
jinja2
